Match saveSelect_getInfo control flow
Invert valid-flag guard to drop a return island, embed the valid
assignment in the guard test to keep the byte live, and re-read the
completion byte per use instead of CSE'ing it. 86.76 -> 95.13%.

+++ src/main/objseq.c
@@ -3204,7 +3204,7 @@ void ObjSeq_ApplyLinkedObjectTransform(u8* obj, u8* seqObj, u8* seq)
 {
     int baseYaw;
     int baseRoll;
-    int basePitch;
+    s16 basePitch;
     f32 baseX;
     f32 baseY;
     f32 baseZ;
